make position short

+++ src/Core/Harmony.Application/Contracts/Repositories/ICardRepository.cs
@@ -8,9 +8,9 @@ namespace Harmony.Application.Contracts.Repositories
         Task<Card> Get(Guid cardId);
         Task<Guid> GetBoardId(Guid cardId);
         Task<Card> Load(Guid cardId);
-        Task<Card> GetByPosition(Guid boardListId, byte position);
-		Task<List<Card>> GetCardsInPositionGreaterThan(Guid boardListId, byte position);
-		Task<List<Card>> GetCardsInPositionGreaterOrEqualThan(Guid boardListId, byte position);
+        Task<Card> GetByPosition(Guid boardListId, short position);
+		Task<List<Card>> GetCardsInPositionGreaterThan(Guid boardListId, short position);
+		Task<List<Card>> GetCardsInPositionGreaterOrEqualThan(Guid boardListId, short position);
 		Task<int> CountCards(Guid listId);
         Task<int> Add(Card card);
 		Task<int> Update(Card card);
