+++ src/Core/Harmony.Application/DTO/BoardListDto.cs
@@ -6,7 +6,7 @@ namespace Harmony.Application.DTO
 	{
 		public Guid Id { get; set; }
 		public string Name { get; set; }
-        public byte Position { get; set; } // position on the board
+        public short Position { get; set; } // position on the board
 		public List<CardDto> Cards { get; set; }
 
 		// helpers for kanban
