+++ src/Core/Harmony.Domain/Entities/BoardList.cs
@@ -11,7 +11,7 @@ namespace Harmony.Domain.Entities
         public Board Board { get; set; }
 		public string UserId { get; set; } // User created the list
 		public Guid BoardId { get; set; }
-        public byte Position { get; set; } // position on the board
+        public short Position { get; set; } // position on the board
         public List<Card> Cards { get; set; }
         public BoardListStatus Status { get; set; }
     }
